Model,Schema: move populate() to schema file
