+++ lib/model.ts
@@ -1,6 +1,6 @@
 import { clone, cloneDeep, defaults, difference, forEach, isArray, isString } from 'lodash';
 
-import { ISchemaSanitizeOptions, ISchemaToObjectOptions, sanitize,
+import { ISchemaSanitizeOptions, ISchemaToObjectOptions, populate, sanitize,
          SchemaMap, SchemaValue, setObjectSanitized, toObject } from './schema';
 
 interface IModel {
@@ -27,95 +27,6 @@ function defineModelProperty(target, key: string, enumerable: boolean) {
 
 export abstract class Model<T> {
   public static _schema: SchemaMap = {};
-
-  // TODO: static
-  public static async getPopulated(obj: object, schemaMap: SchemaMap, path: string[]) {
-    if (path.length === 0) {
-      throw new Error('path is empty');
-    }
-
-    const field = path[0];
-
-    const schemaValue = schemaMap[field];
-    if (!schemaValue) {
-      throw new Error(`field ${field} unknown in schema`);
-    }
-
-    const value = obj[field];
-
-    // final path element?
-    if (path.length === 1) {
-      // throw if not a reference
-      if (schemaValue.type !== 'reference') {
-        throw new Error(`field ${field} is not of type reference`);
-      }
-
-      const collection = schemaValue.collection();
-
-      // do nothing if value is undefined
-      if (value === undefined) {
-        return;
-      }
-
-      // already populated?
-      if (value instanceof collection.model) {
-        return;
-      }
-
-      // not an id?
-      if (!isString(value)) {
-        throw new Error(`field ${field} is not an instance and not an id`);
-      }
-
-      // find by id and set on object
-      const instance = await schemaValue.collection().findById(value);
-      if (!instance) {
-        throw new Error(`id ${value} not found`);
-      }
-      obj[field] = instance;
-    }
-
-    // path length > 1
-    const newPath = path.slice(1, path.length);
-    switch (schemaValue.type) {
-      case 'array':
-        if (value === undefined) {
-          return;
-        }
-        // TODO
-        return;
-
-      case 'model':
-        if (value === undefined) {
-          return;
-        }
-        const instance = value as Model<any>; // TODO: replace any
-        await instance.populate(newPath);
-        return;
-
-      case 'object':
-        if (value === undefined) {
-          return;
-        }
-        await Model.getPopulated(value, schemaValue.definition, newPath);
-        return;
-
-      case 'reference':
-        const collection = schemaValue.collection();
-
-        // populate inside the populated instance
-        if (value instanceof collection.model) {
-          await value.populate(newPath);
-          return;
-        }
-
-        // do nothing if not populated
-        return;
-
-      default:
-        throw new Error(`field ${field} of type ${schemaValue.type} cannot be populated`);
-    }
-  }
 
   /**
    * Attach schema information to the property
@@ -150,7 +61,7 @@ export abstract class Model<T> {
     }
 
     await Promise.all(paths.map(
-      path => Model.getPopulated(this, constructor._schema, isArray(path) ? path : [path]),
+      path => populate(this, constructor._schema, isArray(path) ? path : [path]),
     ));
   }
