+++ lib/reference-array.ts
@@ -10,7 +10,6 @@ export class ReferenceArray<T extends object, TModel extends Model<T>> extends A
     data.forEach(element => this.push(element));
   }
 
-  // TODO: what do we do with undefined? do we overwrite ids with undefined?
   public async populate() {
     const fetchModels = [];
     this.forEach((element, index) => {
@@ -25,6 +24,8 @@ export class ReferenceArray<T extends object, TModel extends Model<T>> extends A
     // fetch models
     const models = await this.collection.findByIds(fetchModels.map(fetchModel => fetchModel.id));
 
+    // throw if an id wasn't found
+    // reason: otherwise we'd replace the id with undefined and this may get persisted to the db
     fetchModels.forEach((fetchModel, index) => {
       if (!models[index]) {
         throw new Error(`id ${fetchModel.id} not found`);
