+++ lib/schema.ts
@@ -84,6 +84,100 @@ export interface ISchemaSanitizeOptions {
   replace?: boolean;
 }
 
+// populate an object
+export async function populate(obj: object, schemaMap: SchemaMap, path: string[]) {
+  if (path.length === 0) {
+    throw new Error('path is empty');
+  }
+
+  const field = path[0];
+
+  const schemaValue = schemaMap[field];
+  if (!schemaValue) {
+    throw new Error(`field ${field} unknown in schema`);
+  }
+
+  const value = obj[field];
+
+  // final path element?
+  if (path.length === 1) {
+    // throw if not a reference
+    if (schemaValue.type !== 'reference') {
+      throw new Error(`field ${field} is not of type reference`);
+    }
+
+    const collection = schemaValue.collection();
+
+    // do nothing if value is undefined
+    if (value === undefined) {
+      return;
+    }
+
+    // already populated?
+    if (value instanceof collection.model) {
+      return;
+    }
+
+    // not an id?
+    if (!isString(value)) {
+      throw new Error(`field ${field} is not an instance and not an id`);
+    }
+
+    // find by id
+    const instance = await schemaValue.collection().findById(value);
+
+    // throw if not found
+    // reason: otherwise we'd replace the id with undefined and this may get persisted to the db
+    if (!instance) {
+      throw new Error(`id ${value} not found`);
+    }
+
+    // set instance
+    obj[field] = instance;
+  }
+
+  // path length > 1
+  const newPath = path.slice(1, path.length);
+  switch (schemaValue.type) {
+    case 'array':
+      if (value === undefined) {
+        return;
+      }
+      // TODO
+      return;
+
+    case 'model':
+      if (value === undefined) {
+        return;
+      }
+      const instance = value as Model<any>; // TODO: replace any
+      await instance.populate(newPath);
+      return;
+
+    case 'object':
+      if (value === undefined) {
+        return;
+      }
+      await populate(value, schemaValue.definition, newPath);
+      return;
+
+    case 'reference':
+      const collection = schemaValue.collection();
+
+      // populate inside the populated instance
+      if (value instanceof collection.model) {
+        await value.populate(newPath);
+        return;
+      }
+
+      // do nothing if not populated
+      return;
+
+    default:
+      throw new Error(`field ${field} of type ${schemaValue.type} cannot be populated`);
+  }
+}
+
 export function sanitize(schemaValue: SchemaValue, data: any, _options?: ISchemaSanitizeOptions) {
   const options = _options || {};
